Don't check parser_t.c against EOF
parser_t.c is a char which may be unsigned on some platforms and
therefore incapable of holding EOF.

This field is only ever set from parser_t.source which is a char array,
which only ever originates from {main,strquote,include} -> parser_pp ->
{init_parser,parser_reset} which all assign it from char arrays.

This means it should never be EOF, so the check can be dropped.

Also update the documentation so it matches.