Initialise getdelim buf pointers to NULL
getdelim can either be passed NULL or a malloc allocated buffer. I am
not sure why on glibc on linux this doesn't outright crash, but it does
crash on my ARM phone.

By initialising to NULL, getdelim correctly allocates a buffer for the
data rather than reallocating a string literal.